// src/interfaces/current.ts
export interface current {
  observation_time: string;
  temperature: number;
  weather_code: number;
  weather_icons?: any;
  weather_descriptions?: any;
  wind_speed: number;
  wind_degree: number;
  wind_dir: string;
  pressure: number;
  precip: number;
  humidity: number;
  cloudcover: number;
  feelslike: any;
  uv_index: number;
  visibility: number;
  is_day: string;
}